% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details}
\alias{details}
\title{Retrieve images and meta-data for a specific record (specimen)}
\usage{
details(Symbiota.ID = 4531213, verbose = TRUE)
}
\arguments{
\item{Symbiota.ID, }{as found in output of \code{records}}

\item{verbose}{logical}
}
\description{
Retrieve images and meta-data for a specific record (specimen)
}
\examples{
\dontrun{
# use function records to download records; then use one of the IDs:
pic <- details(4531213)
# show screenshot of details for this specimen
print(pic$screenshot) # in Viewer (RStudio)
plot(pic$screenshot) # as plot
# Look at one of the images in more detail
print(image_read(pic$urls[2])) # not all links are working in all instances
# Look at meta data; this specimen was collected in 2012 in Washington
pic$meta
}
}
\author{
Franz-Sebastian Krah
}
