% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollections.R
\name{getCollections}
\alias{getCollections}
\title{List of available collections}
\usage{
getCollections()
}
\description{
List of available collections
}
\details{
Get list of avaible collections from the MyCoPortal. For details also see \url{http://mycoportal.org/portal/collections/index.php}
}
\author{
Franz-Sebastian Krah
}
