% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycodist-class.R
\docType{class}
\name{mycodist-class}
\alias{mycodist-class}
\title{An S4 Class to represent query result from the function \link{records}}
\description{
\code{mycodist} holds a records table together with the query meta data and recommended citation
}
\section{Slots}{

\describe{
\item{\code{nr.records}}{A nuneric of the number of records found}

\item{\code{query}}{A list giving the user arguments used in the function}

\item{\code{citation}}{A character string with the recommended citation from the website}

\item{\code{records}}{A data.frame with the query records results}
}}

\seealso{
\code{"\link[=records-class]{mycodist}"}
}
