% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycodist.R
\name{mycodist}
\alias{mycodist}
\title{Create Objects of Class "mycodist"}
\usage{
mycodist(nr.records, citation, query, records)
}
\arguments{
\item{nr.records}{numeric of number of records found}

\item{citation}{character string with recommended citation}

\item{query}{list with user-specified query arguments}

\item{records}{data.frame displaying results from MyCoPortal}
}
\description{
Create objects of class
  \code{"\link[=mycodist-class]{mycodist}"} from meta data
  and a data.frame derived from \code{\link{records}}.
}
