% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickview.R
\name{quickview}
\alias{quickview}
\title{Quickview on images of a record}
\usage{
quickview(Symbiota.ID = 4531213, verbose = TRUE)
}
\arguments{
\item{Symbiota.ID, }{as found in output of \code{records}}

\item{verbose}{logical}
}
\description{
Quickview on images of a record
}
\examples{
\dontrun{
pic <- quickview(4531213)
print(pic) # in Viewer (RStudio)
plot(pic) # as plot
}
}
