% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mycoporta_csv.R
\name{read_mycoportal_csv}
\alias{read_mycoportal_csv}
\title{Read a downloaded folder}
\usage{
read_mycoportal_csv(path)
}
\arguments{
\item{path}{character string, specifying the path to the local CSV file after manual download}
}
\description{
Read a downloaded folder
}
