% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{records}
\alias{records}
\title{Create Objects of Class "records"}
\usage{
records(nr.records, citation, query, records, db)
}
\arguments{
\item{nr.records}{A numeric giving the number of records retrieved}

\item{citation}{A character string with the recommended citation from the website}

\item{query}{A list of the user arguments used}

\item{records}{A data.frame with the query records results}

\item{db}{A character string specifying the database (currently only MyCoPortal)}
}
\description{
Create objects of class
  \code{"\link[=records-class]{mycodist}"} from meta data
  and a data.frame derived from \code{\link{mycoportal}}.
}
\author{
Franz-Sebastian Krah
}
