% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records_ext.R
\name{records_hightax}
\alias{records_hightax}
\title{Retrieve higher taxon records from the MyCoPortal}
\usage{
records_hightax(taxon = "Polyporales", verbose = TRUE,
  screenshot = TRUE, port = 4445L, browserName = "chrome",
  remoteServerAddr = "localhost", sleep = 2)
}
\arguments{
\item{taxon}{character string specifying the taxon name (e.g., species name, family name or higher taxon)}

\item{verbose}{logical}

\item{screenshot}{logical, whether screenshot of results should be displayed in Viewer}

\item{port}{default is 4445L}

\item{browserName}{character string specifying the browser to use, recommended: "chrome"}

\item{remoteServerAddr}{default is "localhost}

\item{sleep}{set time to wait till page is loaded, default: 2 (don't go lower than that)}

\item{browser}{default is "chrome"}
}
\value{
x an object of class "\code{mycodist}" with the following components:
\item{nr.records}{A nuneric of the number of records found}
\item{citation}{A character string with recommended citation}
\item{query}{A list with user-specified query arguments}
\item{records}{A data.frame with results from MyCoPortal}
}
\description{
Retrieve higher taxon records from the MyCoPortal
}
\details{
Interface to the web database MyCoPortal. Query records available by various user specifications.
}
\examples{
\dontrun{
## Query Amanitacae and plot on world map or USA map
poly.dist <- records_hightax(taxon = "polyporales", taxon_type = 2)
recordsmap(poly.dist, mapdatabase = "world", legend = FALSE)
recordsmap(poly.dist, mapdatabase = "state",legend = FALSE)
}
}
\references{
\url{http://mycoportal.org/portal/index.php}
}
