% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stop_docker.R
\name{start_docker}
\alias{start_docker}
\title{Start Docker}
\usage{
start_docker(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\description{
Start Docker
}
\details{
This should run for Unix platforms (e.g., Mac) and Windows. Docker available for download at: https://www.docker.com
}
