% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stop_docker.R
\name{stop_docker}
\alias{stop_docker}
\title{Stop Docker}
\usage{
stop_docker()
}
\description{
Stop Docker
}
\details{
This should run for Unix platforms (e.g., Mac) and Windows. Docker available for download at: https://www.docker.com
}
