% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details}
\alias{details}
\title{Retrieve images and meta-data for a specific record (specimen)}
\usage{
details(Symbiota.ID = 4531213)
}
\arguments{
\item{Symbiota.ID, }{as found in output of \code{records}}
}
\description{
Retrieve images and meta-data for a specific record (specimen)
}
\examples{
\dontrun{
# use function records to download records; then use one of the IDs:
pic <- details(4531213)
# Look at one of the images in more detail
print(image_read(pic$urls[1])) # a rather orangeish specimen
par(mfrow = c(1,3))
plot(image_read(pic$urls[1]))
plot(image_read(pic$urls[2]))
plot(image_read(pic$urls[3]))
# Look at meta data
pic$meta
}
}
\author{
Franz-Sebastian Krah
}
