% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycodist-class.R
\docType{class}
\name{mycodist-class}
\alias{mycodist-class}
\title{An S4 Class to represent query result from the function \link{records}}
\description{
\code{mycodist} holds a records table together with the query meta data and recommended citation
}
\section{Slots}{

\describe{
\item{\code{nr.records}}{A numeric giving the number of records retrieved}

\item{\code{citation}}{A character string with the recommended citation from the website}

\item{\code{query}}{A list of the user arguments used}

\item{\code{records}}{A data.frame with the query records results}
}}

\seealso{
\code{"\link[=mycodist-class]{mycodist}"}
}
\author{
Franz-Sebastian Krah
}
