% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycodist.R
\name{mycodist}
\alias{mycodist}
\title{Create Objects of Class "mycodist"}
\usage{
mycodist(nr.records, citation, query, records)
}
\arguments{
\item{nr.records}{A numeric giving the number of records retrieved}

\item{citation}{A character string with the recommended citation from the website}

\item{query}{A list of the user arguments used}

\item{records}{A data.frame with the query records results}
}
\description{
Create objects of class
  \code{"\link[=mycodist-class]{mycodist}"} from meta data
  and a data.frame derived from \code{\link{records}}.
}
\author{
Franz-Sebastian Krah
}
