% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{records}
\alias{records}
\title{Retrieve records from the MyCoPortal}
\usage{
records(taxon = "Amanita muscaria", country = "", state = "",
  county = "", locality = "", elevation_from = "",
  elevation_to = "", host = "", taxon_type = 1, collection = "all",
  north_lat = "", south_lat = "", west_lon = "", east_lon = "",
  point_lat = "", point_lon = "", radius = "", collector = "",
  collector_num = "", coll_date1 = "", coll_date2 = "",
  syns = TRUE, verbose = TRUE, screenshot = TRUE, port = 4445L,
  browserName = "chrome", remoteServerAddr = "localhost")
}
\arguments{
\item{taxon}{character string specifying the taxon name (e.g., species name, family name or higher taxon)}

\item{country}{character string specifying country, e.g., "USA"}

\item{state}{character string specifying state, e.g., "Massachusetts"}

\item{county}{character string specifying county, e.g., "Worcester"}

\item{locality}{character string specifying locality, e.g., "Harvard Forest"}

\item{elevation_from}{character string, meter, e.g., "1000"}

\item{elevation_to}{character string, meter}

\item{host}{character string specifying host species, e.g., "Betula alba"}

\item{taxon_type}{integer, one of 1 to 5 representing "Family or Scientific Name", "Scientific Name only", "Family Only", "Higher Taxonomy", Common Name"}

\item{collection}{either "all" or a vector or integers with number corresponding to collection. For a list of collections use function \code{getCollections()}}

\item{north_lat}{character string, coordinate e.g., "45"}

\item{south_lat}{character string, coordinate}

\item{west_lon}{character string, coordinate, e.g., "-72"}

\item{east_lon}{character string, coordinate}

\item{point_lat}{character string, coordinate}

\item{point_lon}{character string, coordinate}

\item{radius}{character string, km, e.g., "50"}

\item{collector}{character string specifying collector name}

\item{collector_num}{character string specifying collector number}

\item{coll_date1}{character string specifying collection data from, e.g., "19 August 1926"}

\item{coll_date2}{character string specifying collection data from, e.g., "19 August 2018"}

\item{syns}{logical, if TRUE synonyms from MycoBank and IndexFungorum are searched}

\item{verbose}{logical}

\item{screenshot}{logical, whether screenshot of results should be displayed in Viewer}

\item{port}{default is 4445L}

\item{browserName}{character string specifying the browser to use, recommended: "chrome"}

\item{remoteServerAddr}{default is "localhost}
}
\value{
x an object of class "\code{mycodist}" with the following components:
\item{nr.records}{A nuneric of the number of records found}
\item{citation}{A character string with recommended citation}
\item{query}{A list with user-specified query arguments}
\item{records}{data.frame displaying results from MyCoPortal}
}
\description{
Retrieve records from the MyCoPortal
}
\details{
Interface to the web database MyCoPortal. Query records available by various user specifications.
}
\examples{
\dontrun{
## Query Amanitacae and plot on world map or USA map
am.dist <- records(taxon = "Amanitaceae", taxon_type = 2)
recordsmap(am.dist, mapdatabase = "world", legend = FALSE)
recordsmap(am.dist, mapdatabase = "state",legend = FALSE)
}
}
\references{
\url{http://mycoportal.org/portal/index.php}
}
